refactor remove types
