+++ examples/solana/oyster/index.ts
@@ -1,9 +1,5 @@
 export * as actions from './actions';
 export * from './actions';
-// export * as components from './components';
-// export * from './components';
-// export * as constants from './constants';
-// export * from './constants';
 export * as contexts from './contexts';
 export * from './contexts';
 export * as models from './models';
